+++ src/main/java/br/com/fiap/tech/challenge/grupo/vinte/oito/ChefOnline/core/dtos/cardapio/ItemCardapioDTO.java
@@ -1,12 +1,11 @@
 package br.com.fiap.tech.challenge.grupo.vinte.oito.ChefOnline.core.dtos.cardapio;
 import br.com.fiap.tech.challenge.grupo.vinte.oito.ChefOnline.core.dtos.restaurante.RestauranteDTO;
-import java.math.BigDecimal;
 
 public record ItemCardapioDTO(
     Long id,
     String nome,
     String descricao,
-    BigDecimal preco,
+    Double preco,
     Boolean disponibilidade,
     String foto,
     RestauranteDTO restaurante
